refactor: 이미지 관련 클래스 명 변경
