+++ src/main/java/com/birca/bircabackend/command/birca/application/BirthdayCafeImageFacade.java
@@ -4,8 +4,7 @@ import com.birca.bircabackend.command.birca.domain.BirthdayCafe;
 import com.birca.bircabackend.command.birca.dto.BirthdayCafeImageDeleteRequest;
 import com.birca.bircabackend.command.birca.exception.BirthdayCafeErrorCode;
 import com.birca.bircabackend.common.EntityUtil;
-import com.birca.bircabackend.common.exception.BusinessException;
-import com.birca.bircabackend.common.upload.ImageUploader;
+import com.birca.bircabackend.common.image.ImageRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
@@ -16,27 +15,27 @@ public class BirthdayCafeImageFacade {
 
     private final BirthdayCafeImageValidator birthdayCafeImageValidator;
     private final BirthdayCafeImageService birthdayCafeImageService;
-    private final ImageUploader imageUploader;
+    private final ImageRepository imageRepository;
     private final EntityUtil entityUtil;
 
     public void saveDefaultImage(Long birthdayCafeId, MultipartFile defaultImage) {
         birthdayCafeImageValidator.validateImagesSize(birthdayCafeId);
         entityUtil.getEntity(BirthdayCafe.class, birthdayCafeId, BirthdayCafeErrorCode.NOT_FOUND);
-        String imageUrl = imageUploader.upload(defaultImage);
+        String imageUrl = imageRepository.upload(defaultImage);
         birthdayCafeImageService.saveDefaultImage(birthdayCafeId, imageUrl);
     }
 
     public void updateMainImage(Long birthdayCafeId, MultipartFile mainImage) {
         entityUtil.getEntity(BirthdayCafe.class, birthdayCafeId, BirthdayCafeErrorCode.NOT_FOUND);
-        String imageUrl = imageUploader.upload(mainImage);
+        String imageUrl = imageRepository.upload(mainImage);
         birthdayCafeImageService.updateMainImage(birthdayCafeId, imageUrl)
-                .ifPresent(imageUploader::delete);
+                .ifPresent(imageRepository::delete);
     }
 
     public void delete(Long birthdayCafeId, BirthdayCafeImageDeleteRequest request) {
         entityUtil.getEntity(BirthdayCafe.class, birthdayCafeId, BirthdayCafeErrorCode.NOT_FOUND);
         String imageUrl = request.imageUrl();
         birthdayCafeImageService.delete(imageUrl);
-        imageUploader.delete(imageUrl);
+        imageRepository.delete(imageUrl);
     }
 }
